+++ internal/dataplane/sendconfig/inmemory.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/kong/deck/file"
 	"github.com/sirupsen/logrus"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/deckgen"
@@ -41,18 +40,18 @@ func NewUpdateStrategyInMemory(
 	}
 }
 
-func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState *file.Content) (
+func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState ContentWithHash) (
 	err error,
 	resourceErrors []ResourceError,
 	resourceErrorsParseErr error,
 ) {
 	// Kong will error out if this is set
-	targetState.Info = nil
+	targetState.Content.Info = nil
 
 	// Kong errors out if `null`s are present in `config` of plugins
-	deckgen.CleanUpNullsInPluginConfigs(targetState)
+	deckgen.CleanUpNullsInPluginConfigs(targetState.Content)
 
-	config, err := json.Marshal(targetState)
+	config, err := json.Marshal(targetState.Content)
 	if err != nil {
 		return fmt.Errorf("constructing kong configuration: %w", err), nil, nil
 	}
@@ -68,6 +67,10 @@ func (s UpdateStrategyInMemory) Update(ctx context.Context, targetState *file.Co
 
 func (s UpdateStrategyInMemory) MetricsProtocol() metrics.Protocol {
 	return metrics.ProtocolDBLess
+}
+
+func (s UpdateStrategyInMemory) Type() string {
+	return "InMemory"
 }
 
 // shouldUseFlattenedErrors verifies whether we should pass flatten errors flag to ReloadDeclarativeRawConfig.