+++ internal/dataplane/sendconfig/sendconfig.go
@@ -2,18 +2,20 @@ package sendconfig
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"time"
 
 	"github.com/kong/deck/file"
+	"github.com/kong/go-kong/kong"
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
-	"github.com/kong/kubernetes-ingress-controller/v2/internal/adminapi"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/deckgen"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/failures"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/metrics"
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/util"
 )
 
 // -----------------------------------------------------------------------------
@@ -24,11 +26,22 @@ type UpdateStrategyResolver interface {
 	ResolveUpdateStrategy(client UpdateClient) UpdateStrategy
 }
 
+type AdminAPIClient interface {
+	AdminAPIClient() *kong.Client
+	LastConfigSHA() []byte
+	SetLastConfigSHA([]byte)
+	BaseRootURL() string
+	PluginSchemaStore() *util.PluginSchemaStore
+
+	IsKonnect() bool
+	KonnectRuntimeGroup() string
+}
+
 // PerformUpdate writes `targetContent` to Kong Admin API specified by `kongConfig`.
 func PerformUpdate(
 	ctx context.Context,
 	log logrus.FieldLogger,
-	client *adminapi.Client,
+	client AdminAPIClient,
 	config Config,
 	targetContent *file.Content,
 	promMetrics *metrics.CtrlFuncMetrics,
@@ -58,12 +71,21 @@ func PerformUpdate(
 	}
 
 	updateStrategy := updateStrategyResolver.ResolveUpdateStrategy(client)
+	log = log.WithField("update_strategy", updateStrategy.Type())
 	timeStart := time.Now()
-	err, resourceErrors, resourceErrorsParseErr := updateStrategy.Update(ctx, targetContent)
+	err, resourceErrors, resourceErrorsParseErr := updateStrategy.Update(ctx, ContentWithHash{
+		Content: targetContent,
+		Hash:    newSHA,
+	})
 	duration := time.Since(timeStart)
 
 	metricsProtocol := updateStrategy.MetricsProtocol()
 	if err != nil {
+		// Not pushing metrics in case it's an update skip due to a backoff.
+		if errors.Is(err, ErrUpdateSkippedDueToBackoffStrategy{}) {
+			return nil, []failures.ResourceFailure{}, err
+		}
+
 		resourceFailures := resourceErrorsToResourceFailures(resourceErrors, resourceErrorsParseErr, log)
 		promMetrics.RecordPushFailure(metricsProtocol, duration, client.BaseRootURL(), err)
 		return nil, resourceFailures, err